Join As Drone and drone rebooting will now check cyborg jobbans.

+++ code/modules/mob/living/silicon/robot/drone/drone.dm
@@ -265,7 +265,7 @@
 
 /mob/living/silicon/robot/drone/proc/question(var/client/C)
 	spawn(0)
-		if(!C)	return
+		if(!C || jobban_isbanned(C,"Cyborg"))	return
 		var/response = alert(C, "Someone is attempting to reboot a maintenance drone. Would you like to play as one?", "Maintenance drone reboot", "Yes", "No", "Never for this round.")
 		if(!C || ckey)
 			return
